BUG: Support color images (RGB/RGBA) with DCMTK
To enable support for color images with DMCTK, a static initialization has to
be done by including the header file "dcmtk/dcmimage/diregist.h". This include
was commented out in two locations of the ITK code. The commit reintroduces it.
The color image data are by default organize by plane
(i.e. R1R2R3...G1G2G3...B1B2B3) which is incompatible with the data organization
that ITK expected, which is by pixel (i.e. R1G1B1R2G2B2R3G3B3...). The data has
to be copies and reorganized to be used in ITK. This reorganization is done by
the new function ReorderRGBValues(...) which is templated over the pixel type.
The implementation of this function is based on the function Convert(...)
in DCMTK (in file dcmimage/include/dcmtk/dcmimage/dicoopxt.h). This function
could be called indirectly in DCMTK by getting the pointer to the DCMTK data
with the function getOutputData() available in the type DicomImage, instead
of using the function getInterData(), but replacing this function call
leads to some tests failing and potential non-backward compatibility. Indeed,
this patch only modifies the behavior of itkDCMTKImageIO for color images, that
were previously not working.

Change-Id: I79f17dc2c4fc4f455f48cf9ddcf435863f8282a6