PERF: GDCM & DCMTK CanRead fails very slowly for non-DICOM files.
This commit is a replica of an official ITK bugfix:
https://github.com/InsightSoftwareConsortium/ITK/commit/19fa58feb02bfd45154e811d0e13510787e2adfb#diff-824229b4fd5f76e67acd43acc54fdae2
and subsequent corrections until
https://github.com/Slicer/ITK/commit/35e6f546438557f22e66db25e066499637890214

NOTE: DCMTK and GDCM demonstrate the same behavior with certain
      non-DICOM files that have byte patterns similar to DICOM

Some images take extremely long time to load because
DCMTKImageIO::CanReadFile & GDCMImageIO::CanReadFile takes several
minutes to fail.  This is due to DCMTK's and GDCM's default behavior of
trying to read non-compliant dicom files that do not have the required
DICM header.

Add more extensive testing about the structure of the file to determine
if it looks like a dicom file.  Previous testing only looked to see if the
files without preables had values of 2 or 8 as the first byles of the file,
but that resulted in many false positives.

This implementation looks at all the SOP Instances that start with 2 or 8
to ensure that the proper dicom structure is found.

This resolves #388.